requirements.txt
streamlit==1.32.0
pandas==2.2.1
pyarrow==15.0.0
plotly==5.20.0
Pillow==10.2.0
